+++ lib/enum/ContractType.enum.ts
@@ -1,0 +1,3 @@
+export enum ContractType {
+    Rental = "rental"
+}