+++ lib/interfaces/OnChainRental.interface.ts
@@ -1,0 +1,9 @@
+export default interface IOnChainRental {
+    owner: string;
+    borrower: string;
+    borrowedAt: bigint;
+    publishedAt: bigint;
+    price: bigint;
+    duration: bigint;
+    interest: bigint;
+}