+++ lib/modules/network/getContractAddresses.ts
@@ -1,0 +1,5 @@
+import { MSBlockchain } from "../../services";
+
+export default async function getContractAddresses () {
+    return await MSBlockchain.network.contracts();
+}