+++ lib/modules/web3-utils/toWei.ts
@@ -1,0 +1,5 @@
+export default function toWei(amount: number, decimals: bigint) {
+    let res = BigInt(amount) * (BigInt(10) ** decimals);
+
+    return res
+}