+++ lib/modules/web3/isZeroAddress.ts
@@ -1,0 +1,3 @@
+export async function isZeroAddress (address: string) {
+    return address === "0x0000000000000000000000000000000000000000";
+}