+++ lib/services.ts
@@ -1,0 +1,4 @@
+import { HTTPProxy } from "../plugins/http-proxies";
+import { $BLOCKCHAIN_HOST } from "./env";
+
+export const MSBlockchain = HTTPProxy($BLOCKCHAIN_HOST);