+++ lib/web3/contracts/Rental.contract.ts
@@ -1,0 +1,5 @@
+import createContract from "../../../plugins/web3/contracts";
+
+const RentalContract = (addr: string) => createContract("Rental", addr);
+
+export default RentalContract;