+++ CHANGELOG.md
@@ -8,5 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- [2024-03-23] Install CLI Plugin
+- [2024-03-23] Create 'anfri view rent <anfri-id>' command
+- [2024-03-23] Create `anfri view rent <anfri-id>` command
 - [2024-03-23] Initial Files